Need build tools installed in order to install Buildbot from source
Installing Buildbot for Python 2 probably won't need a compiler
available but installing for Python 3 will. The bundled Twisted isn't
new enough to fix Python 3 compatibility so installing from source will
also install a new Twisted. This makes the image much larger but has to
be done to make it easy to install from source for testing.